scripts/requirements.txt
fastapi
uvicorn
pydantic>=2
pandas
numpy
joblibrequests
